+++ api/xnote.js
@@ -173,7 +173,7 @@ var xnote = class extends ExtensionCommon.ExtensionAPI {
 		}
 
 		function fileExists(file){
-			return file && file.exists() && file.isFile() && file.isReadable();
+			return file && file.exists();
 		}
 
 		function getExistingNoteFile(root, keyId){
@@ -229,7 +229,9 @@ var xnote = class extends ExtensionCommon.ExtensionAPI {
 				async deleteNote(root, keyId){
 					try {
 						var file = getExistingNoteFile(root, keyId);
-						file.remove(false);
+						if(file){
+							file.remove(false);
+						}
 					} catch(e) {
 						throw new ExtensionError(e.message);
 					}
@@ -237,7 +239,6 @@ var xnote = class extends ExtensionCommon.ExtensionAPI {
 				async loadNote(root, keyId){
 					try {
 						var file = getExistingNoteFile(root, keyId);
-
 						if(!file){
 							return false;
 						}
